Choropleth Graph - (Geomap Report in Email)
Choropleth Maps display divided geographical areas or regions that are coloured, shaded or patterned in relation to a data variable. The geographical area may expanse the entire world, or a country, state or even a county. There are many tools and packages available to make a stand alone or static choropleth map using Python. 
References:
https://plotly.com/python/choropleth-maps/
https://towardsdatascience.com/a-complete-guide-to-an-interactive-geographical-map-using-python-f4c5197e23e0